r128-specific parts of sync with 2.4.0-test11-pre5. Rest is coming shortly.

+++ linux/r128_drv.h
@@ -379,7 +379,7 @@ extern int  r128_context_switch_complete(drm_device_t *dev, int new);
 #define R128_MAX_VB_VERTS		(0xffff)
 
 
-#define R128_BASE(reg)		((unsigned long)(dev_priv->mmio->handle))
+#define R128_BASE(reg)		((u32)(dev_priv->mmio->handle))
 #define R128_ADDR(reg)		(R128_BASE(reg) + reg)
 
 #define R128_DEREF(reg)		*(__volatile__ int *)R128_ADDR(reg)
